+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * FastQC: Per base sequence content heatmap can now be clicked to show line graph for single sample
 * FastQC: No longer show adapter contamination datasets with <= 0.1% contamination.
 * New regex help box with examples added to report
+* General Stats table 'floating' header now sorts properly when scrolling
 * Bugfix: MultiQC default_dev template now copies module assets properly
 * Bufgix: General Stats table floating header now resizes properly when page width changes